implement rtb-api and remove the client api, redis, etc
