--- api-clients/dublin/responses/operators.js
@@ -1,3 +1,0 @@
-let operators  = [{ shortName: "bac", completeName: 'Dublin Bus'  },{ shortName: "GAD", completeName: 'Go-Ahead Ireland'}];
-
-module.exports = operators;
