--- api-clients/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-    dublinApi: require('./dublin/dublin-api')
-};
