+++ cmd/index.js
@@ -1,9 +1,9 @@
 module.exports ={
-    getLocalization : require('./get-localization'),
+    getLocation : require('./get-location'),
     getStopByNumber : require('./get-stop-by-number'),
     backMainMenu : require('./back-main-menu'),
     exit: require('./exit'),
     getRouteRealTime: require('./get-route-realtime'),
     getStopsNearMe: require('./get-stops-near-me'),
     help: require('./help')
-}
+};
