+++ real-time-bot.js
@@ -1,9 +1,5 @@
 const bluebird = require('bluebird');
 
-const redisClient = require('./helper/redis-client');
-redisClient.createClient();
-
-
 
 global.Promise = bluebird;
 
@@ -18,7 +14,6 @@ const bot = new TelegramBot(json.authorizationToken, { polling: true })
 bot.onText(/^\/start/, (msg, match) => {
     bot.sendMessage(msg.chat.id, `Welcome to Real Time Bus. This bot will help you to find the best route you need. Enjoy It  ${require('./emoji.js').smillingFace.openMouth}`, { parse_mode: 'HTML' });
     var cmd = require('./cmd/command-match').cmd(msg.text);
-    cmd.redis = redisClient;
     cmd.bot = bot;
     cmd.exec(msg);
 });
@@ -31,7 +26,6 @@ bot.on('callback_query', (callbackQuery) => {
     let arr  = callbackQuery.data.split('|');
     let action = arr[0];
     let cmd = require('./cmd/command-match').cmd(action);
-    cmd.redis = redisClient;
     cmd.bot = bot;
     cmd.exec(callbackQuery);
 });
